@@ -4,29 +4,29 @@
   "private": true,
   "dependencies": {
     "@chat-e2ee/service": "file:../service",
-    "@types/jest": "^29.5.2",
-    "@types/node": "^20.2.5",
-    "@types/react": "^18.2.9",
-    "@types/react-dom": "^18.2.4",
-    "qrcode.react": "^3.1.0",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.5.2",
-    "react-dom": "^16.13.1",
-    "react-icons": "^3.10.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^5.0.1",
-    "typescript-plugin-css-modules": "^5.0.1"
+    "qrcode.react": "^4.0.1",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.4",
+    "react-dom": "^18.3.1",
+    "react-icons": "^5.3.0",
+    "react-router-dom": "^6.26.2"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
+    "react-scripts": "^5.0.1",
+    "typescript-plugin-css-modules": "^5.1.0",
+    "@types/jest": "^29.5.13",
+    "@types/node": "^22.5.5",
+    "@types/react": "^18.3.7",
+    "@types/react-dom": "^18.3.0",
+    "@testing-library/jest-dom": "^6.5.0",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
     "@types/react-router-dom": "^5.3.3",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-promise": "4.1.1",
-    "eslint-plugin-react": "7.31.11",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "html-webpack-plugin": "^5.5.1",
+    "eslint-plugin-jsx-a11y": "^6.10.0",
+    "eslint-plugin-promise": "7.1.0",
+    "eslint-plugin-react": "7.36.1",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "html-webpack-plugin": "^5.6.0",
     "typescript": "^4.9.5"
   },
   "scripts": {
