@@ -24,35 +24,36 @@
     "cors": "^2.8.5",
     "cross-env": "^7.0.3",
     "dotenv": "^16.4.5",
-    "express": "^4.19.2",
+    "express": "^4.21.0",
     "form-data": "^4.0.0",
-    "mongodb": "^6.8.0",
+    "mongodb": "^6.9.0",
     "node-fetch": "^2.7.0",
     "socket.io": "^4.7.5",
     "uuid": "^10.0.0"
   },
   "devDependencies": {
-    "react-bootstrap": "^2.10.4",
-    "ts-node": "^10.9.2",
-    "typescript": "5.5.4",
-    "@commitlint/cli": "^19.4.0",
-    "@commitlint/config-conventional": "^19.2.2",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@commitlint/cli": "^19.5.0",
+    "@commitlint/config-conventional": "^19.5.0",
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
-    "@types/jest": "^29.5.12",
-    "@types/node": "^22.2.0",
+    "@types/jest": "^29.5.13",
+    "@types/node": "^22.5.5",
     "@types/node-fetch": "^2.6.11",
-    "concurrently": "^8.2.2",
-    "eslint": "^9.9.0",
+    "concurrently": "^9.0.1",
+    "eslint": "^9.10.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-jest": "28.8.0",
-    "husky": "^9.1.4",
+    "eslint-plugin-jest": "28.8.3",
+    "husky": "^9.1.6",
     "jest": "^29.7.0",
-    "lint-staged": "^15.2.8",
-    "nodemon": "^3.1.4",
+    "lint-staged": "^15.2.10",
+    "nodemon": "^3.1.6",
     "prettier": "^3.3.3",
+    "react-bootstrap": "^2.10.4",
     "supertest": "^7.0.0",
-    "ts-jest": "^29.2.4"
+    "ts-jest": "^29.2.5",
+    "ts-node": "^10.9.2",
+    "typescript": "5.6.2"
   },
   "husky": {
     "hooks": {
