@@ -5,22 +5,21 @@
   "main": "dist/bundle.js",
   "author": "Mukesh",
   "license": "ISC",
-  "types": "dist/types.d.ts",
+  "types": "dist/types/sdk.d.ts",
   "devDependencies": {
-    "ts-loader": "^9.4.2",
-    "webpack": "^5.82.0",
-    "webpack-cli": "^5.0.2",
+    "ts-loader": "^9.5.1",
+    "webpack": "^5.94.0",
+    "webpack-cli": "^5.1.4",
     "webpack-remove-debug": "^0.1.0"
   },
   "scripts": {
-    "build": "webpack --mode production && npm run dts",
+    "build": "webpack --mode production",
     "watch": "concurrently \"webpack --watch\" \"npm run watch-dts\"",
-    "dts": "../node_modules/typescript/bin/tsc src/public/types.ts --declaration --emitDeclarationOnly --declarationDir ./dist",
     "publish-sdk": "npm publish",
     "test": "jest",
     "watch-dts": "../node_modules/.bin/nodemon --watch src/public/types.ts --exec \"npm run dts\""
   },
   "dependencies": {
-    "socket.io-client": "^4.6.1"
+    "socket.io-client": "^4.7.5"
   }
 }
