@@ -1,6 +1,6 @@
 import makeRequest from './makeRequest';
 
-const getChatLink = async (pin) => {
+const getChatLink = async (pin: string) => {
   return makeRequest(`chat-link/${pin}`, {
     method: 'GET'
   });
